small structural changes
Additions:
-Plane object
-trash.png

Changes:
-Moved component scripts to a Component folder
-Added support for transparent textures
-Added ability to toggle scenes and transforms on and off
-Moved skyboxes to the scenegraph
-Changed skybox such that initializing one with zero parameters makes it not render at all

Fixes:
-

+++ Minimal/Components/ComponentTest4.h
@@ -19,7 +19,7 @@ public:
 	}
 
 	void Update(double deltaTime) override {
-		curTime += deltaTime;
+		curTime += (float)deltaTime;
 		if (curTime > time) {
 			curTime = 0;
 			curDir *= -1.0f;
